+++ jest.config.js
@@ -149,7 +149,7 @@ module.exports = {
   // ],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
-  testRegex: ['./test/src/*', './test/v2.*/*'],
+  testRegex: ['./test/src/*', './test/v2.*/*', './test/v3.*/*'],
 
   // This option allows the use of a custom results processor
   // testResultsProcessor: null,