+++ test/v2.4.0/userMessages.js
@@ -6,7 +6,7 @@ describe('userMessages', () => {
     validator = await AsyncApiValidator.fromSource('./test/schemas/v2.4.0/userMessages.yaml')
   })
 
-  it.only('should validate UserDeleted message', () => {
+  it('should validate UserDeleted message', () => {
     const validate = validator.validateByMessageId('UserDeleted', {
       role: 'admin',
       userName: 'user@gmail.com'